feat: Agents tab shows runs, artifacts, log and settings
`f` opens an agent like a session. The detail leads with whether it needs
you (off, halted and why, the last run's outcome with reason and stderr),
then four sections: Runs (f opens a transcript), Artifacts (f opens a
note's ref), Log, and Settings (edit agent.toml in place, comments kept,
validated before saving).

- Space turns an agent on/off via `enabled`; on also clears a halt.
- `n` opens an attached Claude session in the agent's folder, primed with
  the folder layout and the last failure. Sessions there now show on the
  Sessions tab; only `<agent>/work` stays the agent's.
- `p` says when the queued run will actually start.
- New per-agent events.jsonl: run start/end, poll failures (deduped),
  halts, stale-run reclaims, hub edits.
- Failed runs keep stderr and a longer result.
- API errors reported as subtype `success` with `is_error` now count as
  failures.

+++ lib/src/app/render_state.rs
@@ -40,8 +40,9 @@ pub struct RenderState {
     /// Scroll offset (rows) of the Agents-tab table, written by the
     /// renderer to keep the selected row on screen.
     pub agents_scroll: u16,
-    /// Scroll offset of the agent-detail popup's tick timeline.
-    pub agent_detail_scroll: u16,
+    /// First visible row of the agent detail's open section, written by
+    /// the renderer to keep the cursor on screen.
+    pub agent_detail_scroll: usize,
 }
 
 impl Default for RenderState {
